Componentized IconButton and CanvasImage
