+++ frontend/src/index.css
@@ -67,17 +67,6 @@ button:focus-visible {
   outline: 4px auto -webkit-focus-ring-color;
 }
 
-.poster-title {
-  font-family: "Tiro Devanagari Hindi", serif;
-  font-size: 96px;
-  font-weight: 400;
-  line-height: 1;
-  letter-spacing: 0;
-  color: #48351c;
-  text-align: center;
-  margin: 0 0 12px;
-}
-
 @media (prefers-color-scheme: light) {
   :root {
     color: #213547;
