+++ frontend/src/pages/Canvas/Canvas.css
@@ -1,0 +1,10 @@
+.poster-title {
+  font-family: "Tiro Devanagari Hindi", serif;
+  font-size: 96px;
+  font-weight: 400;
+  line-height: 1;
+  letter-spacing: 0;
+  color: #48351c;
+  text-align: center;
+  margin: 0 0 12px;
+}
